fix: move app preference interface into repo
